+++ src/ipfs-http.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable, Logger } from '@nestjs/common'
-import { create, Options, IPFSHTTPClient } from 'kubo-rpc-client'
 import { CLIENT_OPTIONS } from './constants'
+import { IPFSHTTPClient, Options } from './lib'
 
 @Injectable()
 export class IpfsHttpService {
@@ -8,6 +8,6 @@ export class IpfsHttpService {
     client: IPFSHTTPClient
 
     constructor(@Inject(CLIENT_OPTIONS) private readonly options: Options) {
-        this.client = create(options)
+        this.client = new IPFSHTTPClient(options)
     }
 }