+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { IPFSHTTPClient } from './ipfs'
+export type { AddFileParams, GetFileParams, DhtStatsParams, Options } from './types'