+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+
+### Changed
+
+- Drop `kubo-rpc-client` because ESM/CJS madness
+- Implemented `add`, `get`, `cat`, `stats/dht` manually
+
 ## [0.1.0]
 
 ### Changed